requirements.txt
streamlit
transformers
torch
sentencepiece
geopy
gTTStwilio
pandas
pyperclip
langdetect
plyer
fpdf2
requests
speech_recognition